Fix warnings in tests

+++ server/modules/selva/util/svector.c
@@ -91,7 +91,7 @@ void *SVector_InsertFast(SVector *vec, void *el) {
     while (l <= r) {
         ssize_t m = (l + r) / 2;
 
-        assert(m < vec->vec_last);
+        assert((ssize_t)m < (ssize_t)vec->vec_last);
 
         const int rc = vec->vec_compar((const void **)&el, (const void **)vec_data + m);
         if (rc > 0) {
